allow people to not purge passenger yumrepo
The idea is that people can switch between the different available nginx
repos. in case somebody switched from the passenger one to the upstream
repo we want to purge the passenger repo. However this gets us intro
trouble if the person never switched, but another module declares a
resource with this name.